Add build options for finer-grained control over build flags
GLib has nice build options to allow finer-grained control over its
debug options. GTK should too, because explicit is better than implicit.

Sadly, it is impossible to make our options compatible with the
corresponding GLib options, so I have named them with gtk_ prefixes
instead of glib_ prefixes, and disabled the ability to yield them from
the parent project, which is arguably expected. The problem is GLib's
options are mostly boolean options (except for glib_debug, which is a
feature option) whereas the GTK options have to be feature options. This
is because GTK wants to vary the default behavior of most of its options
based on build type. In contrast, GLib varies only the value of
glib_debug based on build type.

If we want to be compatible with GLib's build options, then we would
need to either (a) change GLib's build options to be feature options
instead of bool options, or (b) not vary the default value of gtk_assert
and gtk_consistency_checks based on build type.